Remove requirement on the GitHub API
This app currently extracts the PR's ID from the webhook payload, and
then makes a request to the GitHub API to fetch the pull request's body.

This won't work for private repos that the govuk-ci account can't access
(for example, DSP repos).

Fortunately it's not really necessary as the payload already contains
the body of the pull request!

This commit:

* removes the GithubPullRequest class and moves its logic into app.rb
* fetches the PR body from the webhook payload instead of the GitHub API
* simplifies the logic in spec/features/app_spec.rb
* makes the response bodies more human-readable to make debugging easier